[#795] Show error when adding duplicate warc file
Show an error when adding a warc file that has the same filename as an existing file in the collection and do not overwrite the existing file.

+++ pywb/manager/manager.py
@@ -123,6 +123,10 @@ directory structure expected by pywb
         full_paths = []
         for filename in warcs:
             filename = os.path.abspath(filename)
+
+            if os.path.exists(os.path.join(self.archive_dir, os.path.basename(filename))):
+                raise IOError(f'Warc {filename} already exists')
+
             shutil.copy2(filename, self.archive_dir)
             full_paths.append(os.path.join(self.archive_dir, filename))
             logging.info('Copied ' + filename + ' to ' + self.archive_dir)
